Rename hooks files with 'use' instead of 'fetch'
