+++ src/components/Explore/DropdownPicker.js
@@ -8,7 +8,7 @@ import { Image } from "react-native";
 // this is a placeholder to get functionality working
 import DropDownPicker from "react-native-dropdown-picker";
 
-import useRemoteSearchResults from "./hooks/fetchSearchResults";
+import useRemoteSearchResults from "./hooks/useRemoteSearchResults";
 import { imageStyles, viewStyles } from "../../styles/explore/explore";
 
 type Props = {
